+++ instrumentation/concurrent_ruby/test/test_helper.rb
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-require 'concurrent/future'
+require 'concurrent'
 
 require 'opentelemetry/sdk'