+++ instrumentation/que/example/run.sh
@@ -1,0 +1,4 @@
+#! /bin/bash
+
+# Start the server, wait a bit, and provide a command line to run the client
+./example.rb & (sleep 5; /bin/bash)