+++ instrumentation/rack/test/opentelemetry/instrumentation/rack/middlewares/tracer_middleware_test.rb
@@ -63,7 +63,7 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
       _(first_span.attributes['http.status_code']).must_equal 200
       _(first_span.attributes['http.target']).must_equal '/'
       _(first_span.attributes['http.url']).must_be_nil
-      _(first_span.name).must_equal '/'
+      _(first_span.name).must_equal 'HTTP GET'
       _(first_span.kind).must_equal :server
     end
 
@@ -80,7 +80,7 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
 
       it 'records the query path' do
         _(first_span.attributes['http.target']).must_equal '/endpoint?query=true'
-        _(first_span.name).must_equal '/endpoint'
+        _(first_span.name).must_equal 'HTTP GET'
       end
     end
 
@@ -91,10 +91,10 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
         it 'does not trace paths listed in the array' do
           Rack::MockRequest.new(rack_builder).get('/ping', env)
 
-          ping_span = finished_spans.find { |s| s.name == '/ping' }
+          ping_span = finished_spans.find { |s| s.attributes['http.target'] == '/ping' }
           _(ping_span).must_be_nil
 
-          root_span = finished_spans.find { |s| s.name == '/' }
+          root_span = finished_spans.find { |s| s.attributes['http.target'] == '/' }
           _(root_span).wont_be_nil
         end
       end
@@ -105,10 +105,10 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
         it 'traces everything' do
           Rack::MockRequest.new(rack_builder).get('/ping', env)
 
-          ping_span = finished_spans.find { |s| s.name == '/ping' }
+          ping_span = finished_spans.find { |s| s.attributes['http.target'] == '/ping' }
           _(ping_span).wont_be_nil
 
-          root_span = finished_spans.find { |s| s.name == '/' }
+          root_span = finished_spans.find { |s| s.attributes['http.target'] == '/' }
           _(root_span).wont_be_nil
         end
       end
@@ -119,10 +119,10 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
         it 'traces everything' do
           Rack::MockRequest.new(rack_builder).get('/ping', env)
 
-          ping_span = finished_spans.find { |s| s.name == '/ping' }
+          ping_span = finished_spans.find { |s| s.attributes['http.target'] == '/ping' }
           _(ping_span).wont_be_nil
 
-          root_span = finished_spans.find { |s| s.name == '/' }
+          root_span = finished_spans.find { |s| s.attributes['http.target'] == '/' }
           _(root_span).wont_be_nil
         end
       end
@@ -138,10 +138,10 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
         it 'does not trace requests in which the callable returns true' do
           Rack::MockRequest.new(rack_builder).get('/assets', env)
 
-          ping_span = finished_spans.find { |s| s.name == '/assets' }
+          ping_span = finished_spans.find { |s| s.attributes['http.target'] == '/assets' }
           _(ping_span).must_be_nil
 
-          root_span = finished_spans.find { |s| s.name == '/' }
+          root_span = finished_spans.find { |s| s.attributes['http.target'] == '/' }
           _(root_span).wont_be_nil
         end
       end
@@ -152,17 +152,23 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
         it 'traces everything' do
           Rack::MockRequest.new(rack_builder).get('/assets', env)
 
-          ping_span = finished_spans.find { |s| s.name == '/assets' }
+          ping_span = finished_spans.find { |s| s.attributes['http.target'] == '/assets' }
           _(ping_span).wont_be_nil
 
-          root_span = finished_spans.find { |s| s.name == '/' }
+          root_span = finished_spans.find { |s| s.attributes['http.target'] == '/' }
           _(root_span).wont_be_nil
         end
       end
     end
 
     describe 'config[:allowed_request_headers]' do
-      let(:env) { Hash('HTTP_FOO_BAR' => 'http foo bar value') }
+      let(:env) do
+        Hash(
+          'CONTENT_LENGTH' => '123',
+          'CONTENT_TYPE' => 'application/json',
+          'HTTP_FOO_BAR' => 'http foo bar value'
+        )
+      end
 
       it 'defaults to nil' do
         _(first_span.attributes['http.request.headers.foo_bar']).must_be_nil
@@ -173,6 +179,22 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
 
         it 'returns attribute' do
           _(first_span.attributes['http.request.headers.foo_bar']).must_equal 'http foo bar value'
+        end
+      end
+
+      describe 'when content-type' do
+        let(:config) { default_config.merge(allowed_request_headers: ['CONTENT_TYPE']) }
+
+        it 'returns attribute' do
+          _(first_span.attributes['http.request.headers.content_type']).must_equal 'application/json'
+        end
+      end
+
+      describe 'when content-length' do
+        let(:config) { default_config.merge(allowed_request_headers: ['CONTENT_LENGTH']) }
+
+        it 'returns attribute' do
+          _(first_span.attributes['http.request.headers.content_length']).must_equal '123'
         end
       end
     end
@@ -245,8 +267,22 @@ describe OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware do
     end
 
     describe 'without quantization' do
-      it 'span.name is uri path' do
+      it 'span.name defaults to low cardinality name HTTP method' do
+        _(first_span.name).must_equal 'HTTP GET'
+        _(first_span.attributes['http.target']).must_equal '/really_long_url'
+      end
+    end
+
+    describe 'with simple quantization' do
+      let(:quantization_example) do
+        ->(url, _env) { url.to_s }
+      end
+
+      let(:config) { default_config.merge(url_quantization: quantization_example) }
+
+      it 'sets the span.name to the full path' do
         _(first_span.name).must_equal '/really_long_url'
+        _(first_span.attributes['http.target']).must_equal '/really_long_url'
       end
     end