+++ sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History: opentelemetry-sdk
 
+### v1.0.1 / 2021-10-29
+
+* FIXED: Add unexpected error handlign in BSP and OTLP exporter (#995)
+
+### v1.0.0 / 2021-09-29
+
+* (No significant changes)
+
 ### v1.0.0.rc3 / 2021-08-12
 
 * BREAKING CHANGE: Remove optional parent_context from in_span